ameliorate halved range of away time comparison due to hitherto 32-bit signed cast and allow detection of > 24 day idle

+++ win32/MainWindow.cpp
@@ -1158,7 +1158,7 @@ void MainWindow::updateStatus() {
 
 	if(SETTING(AWAY_IDLE)) {
 		LASTINPUTINFO info = { sizeof(LASTINPUTINFO) };
-		if((::GetLastInputInfo(&info) && static_cast<int>(::GetTickCount() - info.dwTime) > SETTING(AWAY_IDLE) * 60 * 1000) ^ awayIdle) {
+		if((::GetLastInputInfo(&info) && static_cast<int64_t>(::GetTickCount() - info.dwTime) > SETTING(AWAY_IDLE) * 60 * 1000) ^ awayIdle) {
 			awayIdle = !awayIdle;
 			awayIdle ? Util::incAway() : Util::decAway();
 		}
